+++ src/components/News/_NewsForm.vue
@@ -67,7 +67,7 @@ export default {
             .then(async (downloadURL) => {
               _this.imageURL = downloadURL;
             })
-            .then(function() {
+            .then(function () {
               Newsdb.add({
                 title: _this.title,
                 description: _this.description,
@@ -75,13 +75,13 @@ export default {
                 imageURL: _this.imageURL,
               });
             })
-            .then(function() {
+            .then(function () {
               _this.title = "";
               _this.description = "";
               _this.date = "";
               _this.img = null;
             })
-            .catch(function() {
+            .catch(function () {
               console.log("err");
             });
         });
@@ -89,7 +89,7 @@ export default {
         alert("記入していない部分があります");
       }
     },
-    setImage: function(e) {
+    setImage: function (e) {
       e.preventDefault();
       var file = e.target.files;
       console.log(file);
